Simplify (YAGNI): inline WidgetedComponent mixin
Inheriting from `Grids::WidgetComponent` should suffice for users
for now.
