--- modules/grids/app/components/grids/widgeted_component.rb
@@ -1,53 +1,0 @@
-# frozen_string_literal: true
-
-#-- copyright
-# OpenProject is an open source project management software.
-# Copyright (C) the OpenProject GmbH
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License version 3.
-#
-# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-# Copyright (C) 2006-2013 Jean-Philippe Lang
-# Copyright (C) 2010-2013 the ChiliProject Team
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-# See COPYRIGHT and LICENSE files for more details.
-#++
-
-module Grids
-  module WidgetedComponent
-    extend ActiveSupport::Concern
-    include OpPrimer::ComponentHelpers
-    include OpTurbo::Streamable
-
-    included do
-      option :current_user, default: -> { User.current }
-    end
-
-    def title
-      raise NotImplementedError
-    end
-
-    def widget_wrapper(**, &)
-      component_wrapper tag: "turbo-frame", **, &
-    end
-
-    def wrapper_arguments
-      {}
-    end
-  end
-end
